Rework AOSPA boosting configuration
This changes and updates the logic of how our
custom boosts are applied.
For upstream compatibility these changes are not
applied to Qualcomm's boosts.

The first change is to remove the booleans which
need to be overlayed by the device in order to activate
a boost. Instead, we activate a boost if the parameter
array is non-empty.

The second change is to generalize the weak and
strong parameter definitions for Keypress Boost.

This change also bundles a minor improvement for
Bootup Boost and comment improvements.

Carbon edit: Remove FP boost for now, code location has changed

Change-Id: I550e8b264dce248d152bd754d151aaebd12c3a76
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: Nick <nickgunale@gmail.com>
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>